SERVER-14192 Add circle / polygon containment and intersection test

+++ src/mongo/db/geo/r2_region_coverer.cpp
@@ -175,7 +175,7 @@ namespace mongo {
             // at the same level, we prefer the cells with the smallest number of
             // intersecting children.  Finally, we prefer cells that have the smallest
             // number of children that cannot be refined any further.
-            int priority = -((((candidate->cell.getBits() << 4)
+            int priority = -(((((int)candidate->cell.getBits() << 4)
                               + candidate->numChildren) << 4)
                               + numTerminals);
             _candidateQueue->push(make_pair(priority, candidate)); // queue owns candidate
